days-since-app-backend/requirements.txt
aiosql==10.1
blinker==1.8.2
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
nltk==3.8.1
psycopg==3.1.19
psycopg-pool==3.2.2
python-dotenv==1.0.1
typing_extensions==4.12.0
Werkzeug==3.0.3
